+++ app/pages/game.py
@@ -1,9 +1,9 @@
-from ...framework.core import RuntimeUnit
-from ...framework.io import (
+from ...lega.core import RuntimeUnit
+from ...lega.io import (
     color_theme, font_theme, resolution_info,
     StringFileLikeObject,
 )
-from ...framework.ui import (
+from ...lega.ui import (
     DynamicLabel, LabelButton, PageWithButtons
 )
 from pygame import Surface