+++ app/pages/logo.py
@@ -1,26 +1,26 @@
-from framework.core import RuntimeUnit
-from framework.ui import StaticLabel, Page
-from framework.an.fade_out import FadeOut
+from lega.core import RuntimeUnit
+from lega.ui import StaticLabel, Page
+from lega.an.fade_out import FadeOut
 
 from config import font_theme
 
-from framework.globe import scrmgr
+from lega.globe import scrmgr
 
 
 class Logo(Page):
-    def __init__(self, rtu: RuntimeUnit):
-        Page.__init__(self, rtu)
+    def __init__(self):
+        Page.__init__(self)
 
         self.register_controls(
             # construct "presented by" label
             StaticLabel(
                 content="presented by", font=font_theme.logo, size=10,
-                centerx=scrmgr.x_mean, bottom=scrmgr.y_mean,
+                centerx=scrmgr.center.x, bottom=scrmgr.center.y,
             ),
             # construct "WERTech" label
             StaticLabel(
                 content="WERTech", font=font_theme.logo, size=20,
-                centerx=scrmgr.x_mean, top=scrmgr.y_mean,
+                centerx=scrmgr.center.x, top=scrmgr.center.y,
             ),
         )
 
@@ -28,4 +28,4 @@ class Logo(Page):
         self.draw_and_flip()
 
         # full-screen fade-out
-        FadeOut(self.rtu, scrmgr.screen, click_optional=True, count_down=1000).play()
+        FadeOut(scrmgr.screen, click_optional=True, count_down=1000).play()