+++ main.py
@@ -1,14 +1,11 @@
 import pygame
 
-import framework.core as core
+import lega.core as core
 import app.pages
 
 if __name__ == "__main__":
     pygame.init()
 
-    rtu = core.RuntimeUnit()
-
-    app.pages.Logo(rtu).run()
-
+    app.pages.Logo().run()
     while True:
-        app.pages.Home(rtu).run()
+        app.pages.Home().run()